Refactor: Decouple styleguide data loading from webpack loader
