+++ loaders/styleguide-loader.js
@@ -1,27 +1,9 @@
 'use strict';
 
-const pick = require('lodash/pick');
 const commonDir = require('common-dir');
 const generate = require('escodegen').generate;
 const toAst = require('to-ast');
-const getAllComponentsWithExamples = require('./utils/getAllComponentsWithExamples');
-const getAllContentPages = require('./utils/getAllContentPages');
-const getComponentFilesFromSections = require('./utils/getComponentFilesFromSections');
-const getSections = require('./utils/getSections');
-const filterComponentsWithExample = require('./utils/filterComponentsWithExample');
-
-/* eslint-disable no-console */
-
-// Config options that should be passed to the client
-const CLIENT_CONFIG_OPTIONS = [
-	'title',
-	'highlightTheme',
-	'showCode',
-	'showSidebar',
-	'previewDelay',
-	'theme',
-	'styles',
-];
+const getStyleguide = require('./utils/getStyleguide');
 
 module.exports = function() {};
 module.exports.pitch = function() {
@@ -31,45 +13,26 @@ module.exports.pitch = function() {
 	}
 
 	const config = this._styleguidist;
+	const styleguide = getStyleguide(config);
 
-	let sections = getSections(config.sections, config);
-	if (config.skipComponentsWithoutExample) {
-		sections = filterComponentsWithExample(sections);
-	}
-
-	const allComponentFiles = getComponentFilesFromSections(config.sections, config.configDir);
-	const allContentPages = getAllContentPages(sections);
-	const allComponentsWithExamples = getAllComponentsWithExamples(sections);
-
-	const welcomeScreen = {
-		// Nothing to show in the style guide
-		components: allContentPages.length === 0 && allComponentFiles.length === 0,
-		// All component have no example files
-		examples: allContentPages.length === 0 && allComponentFiles.length > 0 && allComponentsWithExamples.length === 0,
-	};
-
+	/* eslint-disable no-console */
 	/* istanbul ignore if */
 	if (config.verbose) {
 		console.log();
 		console.log('Loading components:');
-		console.log(allComponentFiles.join('\n'));
+		console.log(styleguide.allComponentFiles.join('\n'));
 		console.log();
 	}
+	/* eslint-enable */
 
 	// Setup Webpack context dependencies to enable hot reload when adding new files
 	if (config.contextDependencies) {
 		config.contextDependencies.forEach(dir => this.addContextDependency(dir));
 	}
-	else if (allComponentFiles.length > 0) {
+	else if (styleguide.allComponentFiles.length > 0) {
 		// Use common parent directory of all components as a context
-		this.addContextDependency(commonDir(allComponentFiles));
+		this.addContextDependency(commonDir(styleguide.allComponentFiles));
 	}
-
-	const styleguide = {
-		config: pick(config, CLIENT_CONFIG_OPTIONS),
-		welcomeScreen,
-		sections,
-	};
 
 	return `
 if (module.hot) {
